usm: Move storeUSMStats to a common file
Moving a method that is OS agnostic, but used by multiple OS to a shared file without buildtags
